+++ test/mustache_test.rb
@@ -196,10 +196,13 @@ end_section
 
   def test_classify
     assert_equal 'TemplatePartial', Mustache.classify('template_partial')
+    assert_equal 'Admin::TemplatePartial', Mustache.classify('admin/template_partial')
   end
 
   def test_underscore
     assert_equal 'template_partial', Mustache.underscore('TemplatePartial')
+    assert_equal 'admin/template_partial', Mustache.underscore('Admin::TemplatePartial')
+    assert_equal 'views/in/sub/directories', Mustache.underscore('Views::In::Sub::Directories')
   end
 
   def test_anon_subclass_underscore
@@ -208,7 +211,12 @@ end_section
   end
 
   def test_namespaced_underscore
-    assert_equal 'stat_stuff', Mustache.underscore('Views::StatStuff')
+    Object.const_set(:Views, Class.new)
+    klass = Class.new(Mustache)
+    klass.view_namespace = Views
+    assert_equal 'stat_stuff', klass.underscore('Views::StatStuff')
+
+    assert_equal 'views/stat_stuff', Mustache.underscore('Views::StatStuff')
   end
 
   def test_render
@@ -231,6 +239,24 @@ data
     assert_equal expected, Mustache.render(template, :stage => 'production',
                                                      :server => 'example.com',
                                                      :deploy_to => '/var/www/example.com' )
+  end
+
+  def test_render_from_symbol
+    expected = <<-data
+<VirtualHost *>
+  ServerName example.com
+  DocumentRoot /var/www/example.com
+  RailsEnv production
+</VirtualHost>
+data
+    old_path, Mustache.template_path = Mustache.template_path, File.dirname(__FILE__) + "/fixtures"
+    old_extension, Mustache.template_extension = Mustache.template_extension, "conf"
+
+    assert_equal expected, Mustache.render(:passenger, :stage => 'production',
+                                                       :server => 'example.com',
+                                                       :deploy_to => '/var/www/example.com' )
+
+    Mustache.template_path, Mustache.template_extension = old_path, old_extension
   end
 
   def test_doesnt_execute_what_it_doesnt_need_to
@@ -365,6 +391,17 @@ data
     assert_equal 1, view.calls
   end
 
+  def test_sections_which_refer_to_unary_method_call_them_as_proc
+    kls = Class.new(Mustache) do
+      def unary_method(arg)
+        "(#{arg})"
+      end
+    end
+
+    str = kls.render("{{#unary_method}}test{{/unary_method}}")
+    assert_equal "(test)", str
+  end
+
   def test_lots_of_staches
     template = "{{{{foo}}}}"
 
@@ -469,6 +506,64 @@ def indent
   puts :indented!
 end
 template
+  end
+
+  def test_struct
+    person = Struct.new(:name, :age)
+    view = Mustache.new
+    view[:person] = person.new('Marvin', 25)
+    view.template = '{{#person}}{{name}} is {{age}}{{/person}}'
+    assert_equal 'Marvin is 25', view.render
+  end
+
+  def test_custom_escaping
+    view = Class.new(Mustache) do
+      def escapeHTML(str)
+        "pong"
+      end
+    end
+
+    assert_equal 'pong', view.render("{{thing}}", :thing => "nothing")
+    assert_equal 'nothing', Mustache.render("{{thing}}", :thing => "nothing")
+  end
+
+  def test_implicit_iterator
+    view = Mustache.new
+    view.template = "{{#people}}* {{.}}\n{{/people}}"
+    view[:people] = %w( Chris Mark Scott )
+
+    assert_equal <<text, view.render
+* Chris
+* Mark
+* Scott
+text
+  end
+
+  def test_unescaped_implicit_iterator
+    view = Mustache.new
+    view.template = "{{#people}}* {{{.}}}\n{{/people}}"
+    view[:people] = %w( Chris Mark Scott )
+
+    assert_equal <<text, view.render
+* Chris
+* Mark
+* Scott
+text
+  end
+
+  def test_dot_notation
+    assert_equal <<-text, DotNotation.render
+* Chris Firescythe
+* 24
+* Cincinnati, OH
+* Cincinnati, OH
+* Cincinnati, OH
+* Cincinnati, OH
+* Normal
+
+* Chris Firescythe
+* Cincinnati, OH
+text
   end
 
   def test_inherited_attributes